+++ src/components/rules/rules.ts
@@ -3,7 +3,7 @@ import WebComponent, { createElementFromString } from '../../common/WebComponent
 
 class RulesComponent implements WebComponent {
   render(): Element {
-    return createElementFromString(`<main>
+    return createElementFromString(`<section>
       <div class="container">
         <p class="text text--first">The game consists of a number of rounds. In each round, a player gets three rolls of the dice,
             although they can choose to end their turn after one or two rolls. In each turn a player may throw the dice
@@ -62,7 +62,7 @@ class RulesComponent implements WebComponent {
         <p class="text">The winner is the player with the highest total.</p>
         <p class="text text--last">Enjoy the game!</p>
       </div>
-    </main>`);
+    </section>`);
   }
 
   setup(): void {